fix bugs, wrap everything up

+++ contracts/src/governance/extensions/governor_settings/impls.rs
@@ -8,11 +8,11 @@ use crate::{
 use openbrush::traits::Storage;
 
 pub trait GovernorSettingsImpl: Storage<Data> + GovernorSettingsInternal {
-    fn set_voting_delay(&mut self, new_voting_delay: u128) -> Result<(), GovernanceError> {
+    fn set_voting_delay(&mut self, new_voting_delay: u64) -> Result<(), GovernanceError> {
         self._set_voting_delay(new_voting_delay)
     }
 
-    fn set_voting_period(&mut self, new_voting_period: u128) -> Result<(), GovernanceError> {
+    fn set_voting_period(&mut self, new_voting_period: u64) -> Result<(), GovernanceError> {
         self._set_voting_period(new_voting_period)
     }
 
@@ -20,11 +20,11 @@ pub trait GovernorSettingsImpl: Storage<Data> + GovernorSettingsInternal {
         self._set_proposal_threshold(new_proposal_threshold)
     }
 
-    fn voting_delay(&self) -> u128 {
+    fn voting_delay(&self) -> u64 {
         self.data().voting_delay.get_or_default()
     }
 
-    fn voting_period(&self) -> u128 {
+    fn voting_period(&self) -> u64 {
         self.data().voting_period.get_or_default()
     }
